# specs/008-dashboard-ui-with/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tools Sentiment Analysis API
  description: API endpoints for tool-specific sentiment tracking, comparison, and time series analysis
  version: 1.0.0
  contact:
    name: SentimentAgent API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.sentimentagent.com/api/v1
    description: Production server
paths:
...
  /admin/tools/pending:
    get:
      summary: List pending tool approvals
      description: Admin endpoint to view auto-detected tools awaiting approval (FR-014)
      operationId: listPendingTools
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of pending tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending_tools:
                    type: array
                    items:
                      $ref: '#/components/schemas/PendingTool'
        '401':
          description: Unauthorized - admin auth required
  /admin/tools/{tool_id}/approve:
    post:
      summary: Approve a pending tool
      description: Admin action to approve auto-detected tool (FR-012)
      operationId: approveTool
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tool approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITool'
        '404':
          description: Tool not found
        '400':
          description: Tool not in pending status
  /admin/tools/{tool_id}/reject:
    post:
      summary: Reject a pending tool
      description: Admin action to reject auto-detected tool (FR-012)
      operationId: rejectTool
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Rejection reason
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "False positive - not an AI tool"
      responses:
        '200':
          description: Tool rejected successfully
        '404':
          description: Tool not found
components:
  schemas:
    AITool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "GitHub Copilot"
        vendor:
          type: string
          example: "GitHub/Microsoft"
        category:
          type: string
          enum: [code_assistant, chat_bot, other]
        aliases:
          type: array
          items:
            type: string
          example: ["copilot", "co-pilot", "github copilot"]
        status:
          type: string
          enum: [pending, approved, rejected]
        created_at:
          type: string
          format: date-time
...
    PendingTool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        aliases:
          type: array
          items:
            type: string
        mention_count:
          type: integer
          description: Number of mentions that triggered detection
        first_detected_at:
          type: string
          format: date-time
        sample_mentions:
          type: array
          description: Example posts containing this tool
          items:
            type: object
            properties:
              content_excerpt:
                type: string
              subreddit:
                type: string
              confidence:
                type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT